Signal value panel is added

+++ src/dataVisualizer/DataPanelTimeBased.java
@@ -53,7 +53,7 @@ public class DataPanelTimeBased extends DataPanel {
                   for (int i = 0; i < dataChannelList.size(); i++)
                   {
                       DataChannelListItem dcli = dataChannelList.get(i);
-                      double val = dcli.getDouble(cursors[0].hPos);
+                      double val = dcli.getDouble((int)cursors[0].hPos);
                       String unit = dcli.ch.getUnit();
                       if (!unit.isEmpty())
                           unit = " " + unit;
